+++ pages/pastconferences/pastspeakers/2019/galiya-warrier.js
@@ -14,21 +14,21 @@ import {
 
 library.add(faInstagram, faTwitter, faGithub, faLinkedin);
 
-export default function JessicaSalisbury() {
+export default function GaliyaWarrier() {
   return (
     <Layout>
       <Head>
-        <title>Jessica Salisbury</title>
+        <title>Galiya Warrier</title>
       </Head>
       <Header
-        title={'Jessica Salisbury'}
+        title={'Galiya Warrier'}
         banner="/static/banners/information/thencc.jpg"
       />
       <section>
         <PastSpeakersButton />
         <img
-          src="/static/speakers/2019/Jessica_Salisbury.png"
-          alt="Picture of Jessica Salisbury"
+          src="/static/speakers/2019/Galiya_Warrier.png"
+          alt="Picture of Galiya Warrier"
           className="speakerpic"
         />
         <div>
@@ -40,75 +40,32 @@ export default function JessicaSalisbury() {
           </ul>
         </div>
         <br />
-        <h2 tabIndex="0">The Power of the Feedback Loop</h2>
+        <h2 tabIndex="0">
+          Deep Learning in the world of little ponies
+        </h2>
         <p tabIndex="0">
-          From how the body knows when to stop growing to saving lives
-          from speeding cars, the feedback loop is a mechanism that
-          secretly underlies every corner of our existence. The
-          concept spans schools of thought in psychology, economics,
-          biology and neuroscience and is now emerging in the
-          intersection of technology and behavioural science.
-        </p>
-        <p tabIndex="0">
-          As technologists we are always looking to build products
-          that are engaging and effective, and as people we strive to
-          reach our full personal and working potential. Could the
-          feedback loop be the piece we&apos;re all missing?
-        </p>
-        <p tabIndex="0">
-          In this talk we&apos;ll dive into the structure of the
-          feedback loop, some impressive examples of how powerful they
-          can be and we&apos;ll explore how you can harness their
-          power in your work and personal life.
+          In this talk, we will discuss computer vision, one of the
+          most common real-world applications of machine learning. We
+          will deep dive into various state-of-the-art concepts around
+          building custom image classifiers - application of deep
+          neural networks, specifically convolutional neural networks
+          and transfer learning. We will demonstrate how those
+          approaches could be used to create your own image classifier
+          to recognise the characters of "My Little Pony" TV Series
+          [or Pokemon, or Superheroes, or your custom images].
         </p>
       </section>
 
       <section>
-        <h2 tabIndex="0">About Jessica</h2>
+        <h2 tabIndex="0">About Galiya</h2>
         <p tabIndex="0">
-          <b>
-            Behavioural Psychologist and Money Coaching Project
-            Manager at Tully.
-          </b>
+          <b>Cloud Solution Architect (AA and AI), Microsoft</b>
         </p>
         <p tabIndex="0">
-          Jess is a Behavioural Psychologist and the Project Manager
-          of Money Coaching at Tully, a business created to help
-          people who are worried about money. Jess leads on Tully’s
-          Money Coaching product which uses Open Banking data to
-          engage consumers in positive financial decision making to
-          improve their financial wellbeing. From helping those in
-          debt manage their money to providing early intervention for
-          those displaying detrimental patterns of behaviour, Jess is
-          using her academic background in psychology to create highly
-          personalised, data led Money Coaching that drives long term,
-          financial behavioural change for Tully customers.
+          I&apos;m a Cloud Solution Architect at Microsoft, where I
+          help enterprise customers adopt Advanced Analytics and
+          Artifical Intelligence services on Microsoft Azure cloud.
         </p>
-      </section>
-
-      <section>
-        <h2 tabIndex="0">Video from 2019</h2>
-        <div className="video-frame">
-          <iframe
-            width="560"
-            height="315"
-            src="https://www.youtube.com/embed/bxz6LSOclXQ"
-            title="YouTube video player"
-            frameborder="0"
-            allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture"
-            allowfullscreen
-          ></iframe>
-        </div>
-        <div className="center-link">
-          <a
-            target="_blank"
-            rel="noreferrer"
-            aria-label="Jessica Salisbury: Youtube of Talk"
-            href="https://youtu.be/bxz6LSOclXQ"
-          >
-            Link to Youtube Recording
-          </a>
-        </div>
       </section>
 
       <section>
@@ -117,8 +74,8 @@ export default function JessicaSalisbury() {
           <a
             target="_blank"
             rel="noreferrer"
-            aria-label="Jessica Salisbury: Twitter"
-            href="https://twitter.com/JessSalisburyy"
+            aria-label="Galiya Warrier: Twitter"
+            href="https://twitter.com/galiyawarrier"
           >
             Twitter
           </a>
@@ -127,8 +84,8 @@ export default function JessicaSalisbury() {
           <a
             target="_blank"
             rel="noreferrer"
-            aria-label="Jessica Salisbury: LinkedIn"
-            href="https://www.linkedin.com/in/jessica-salisbury-2a0920102/"
+            aria-label="Galiya Warrier: LinkedIn"
+            href="https://www.linkedin.com/in/galiyawarrier/"
           >
             LinkedIn
           </a>
