+++ pages/pastconferences/pastspeakers/2019/ian-cooper.js
@@ -14,75 +14,72 @@ import {
 
 library.add(faInstagram, faTwitter, faGithub, faLinkedin);
 
-export default function JessicaSalisbury() {
+export default function IanCooper() {
   return (
     <Layout>
       <Head>
-        <title>Jessica Salisbury</title>
+        <title>Ian Cooper</title>
       </Head>
       <Header
-        title={'Jessica Salisbury'}
+        title={'Ian Cooper'}
         banner="/static/banners/information/thencc.jpg"
       />
       <section>
         <PastSpeakersButton />
         <img
-          src="/static/speakers/2019/Jessica_Salisbury.png"
-          alt="Picture of Jessica Salisbury"
+          src="/static/speakers/2019/Ian_Cooper.png"
+          alt="Picture of Ian Cooper"
           className="speakerpic"
         />
         <div>
           <ul>
-            <b>Pronouns: </b> She/Her
+            <b>Pronouns: </b> He/Him
           </ul>
           <ul>
             <b>Talk Length: </b> 60 minute
           </ul>
         </div>
         <br />
-        <h2 tabIndex="0">The Power of the Feedback Loop</h2>
+        <h2 tabIndex="0">How to Escape The Distributed Monolith</h2>
         <p tabIndex="0">
-          From how the body knows when to stop growing to saving lives
-          from speeding cars, the feedback loop is a mechanism that
-          secretly underlies every corner of our existence. The
-          concept spans schools of thought in psychology, economics,
-          biology and neuroscience and is now emerging in the
-          intersection of technology and behavioural science.
+          Microservices were all the rage, so you broke up your
+          monolith. The services talk to each other by gRPC, you use a
+          service mesh to route and load balance, and provide
+          reliability oriented computing, you are fully buzzword
+          compliant.
         </p>
+        <p tabIndex="0">Yet something seems to be wrong.</p>
         <p tabIndex="0">
-          As technologists we are always looking to build products
-          that are engaging and effective, and as people we strive to
-          reach our full personal and working potential. Could the
-          feedback loop be the piece we&apos;re all missing?
+          You can&apos;t easily release software from one team,
+          without coordinating with teams creating other. Testing has
+          to be end-to-end to flush out problems or risk your team
+          creating breaking changes for another team. Your
+          'heavy-lifters', whether you call the principals or
+          architects seem to spend all their time on Docker, K8s,
+          Istio and a whole slew of infrastructure technologies. Your
+          system won&apos;t run without them, and you feel locked in.
         </p>
+        <p tabIndex="0">What happened?</p>
         <p tabIndex="0">
-          In this talk we&apos;ll dive into the structure of the
-          feedback loop, some impressive examples of how powerful they
-          can be and we&apos;ll explore how you can harness their
-          power in your work and personal life.
+          In this talk we look at the emerging world of "smart proxies
+          and dumb endpoints" and ask whatever happened to the vision
+          of "smart endpoints and dumb pipes", and what you can do to
+          change course and deliver on the original promises of
+          microservices to allow your teams to release frequently and
+          independently of each other. And become masters of your tech
+          stack, not its servants.
         </p>
       </section>
 
       <section>
-        <h2 tabIndex="0">About Jessica</h2>
+        <h2 tabIndex="0">About Ian</h2>
         <p tabIndex="0">
-          <b>
-            Behavioural Psychologist and Money Coaching Project
-            Manager at Tully.
-          </b>
+          <b>Coding architect, pierced, bearded, tattooed</b>
         </p>
         <p tabIndex="0">
-          Jess is a Behavioural Psychologist and the Project Manager
-          of Money Coaching at Tully, a business created to help
-          people who are worried about money. Jess leads on Tully’s
-          Money Coaching product which uses Open Banking data to
-          engage consumers in positive financial decision making to
-          improve their financial wellbeing. From helping those in
-          debt manage their money to providing early intervention for
-          those displaying detrimental patterns of behaviour, Jess is
-          using her academic background in psychology to create highly
-          personalised, data led Money Coaching that drives long term,
-          financial behavioural change for Tully customers.
+          Polyglot Coding Architect in London, founder of #ldnug,
+          speaker, tabletop gamer, geek. Tattooed, pierced, and
+          bearded. The 'guv' on @BrighterCommand
         </p>
       </section>
 
@@ -92,7 +89,7 @@ export default function JessicaSalisbury() {
           <iframe
             width="560"
             height="315"
-            src="https://www.youtube.com/embed/bxz6LSOclXQ"
+            src="https://www.youtube.com/embed/CZ_PlBAXUmg"
             title="YouTube video player"
             frameborder="0"
             allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture"
@@ -103,8 +100,8 @@ export default function JessicaSalisbury() {
           <a
             target="_blank"
             rel="noreferrer"
-            aria-label="Jessica Salisbury: Youtube of Talk"
-            href="https://youtu.be/bxz6LSOclXQ"
+            aria-label="Ian Cooper: Youtube of Talk"
+            href="https://youtu.be/CZ_PlBAXUmg"
           >
             Link to Youtube Recording
           </a>
@@ -117,8 +114,8 @@ export default function JessicaSalisbury() {
           <a
             target="_blank"
             rel="noreferrer"
-            aria-label="Jessica Salisbury: Twitter"
-            href="https://twitter.com/JessSalisburyy"
+            aria-label="Ian Cooper: Twitter"
+            href="https://twitter.com/icooper"
           >
             Twitter
           </a>
@@ -127,8 +124,8 @@ export default function JessicaSalisbury() {
           <a
             target="_blank"
             rel="noreferrer"
-            aria-label="Jessica Salisbury: LinkedIn"
-            href="https://www.linkedin.com/in/jessica-salisbury-2a0920102/"
+            aria-label="Ian Cooper: LinkedIn"
+            href="https://www.linkedin.com/in/ian-cooper-2b059b/"
           >
             LinkedIn
           </a>
