+++ pages/pastconferences/pastspeakers/2019/ian-johnson.js
@@ -14,101 +14,74 @@ import {
 
 library.add(faInstagram, faTwitter, faGithub, faLinkedin);
 
-export default function JessicaSalisbury() {
+export default function IanJohnson() {
   return (
     <Layout>
       <Head>
-        <title>Jessica Salisbury</title>
+        <title>Ian Johnson</title>
       </Head>
       <Header
-        title={'Jessica Salisbury'}
+        title={'Ian Johnson'}
         banner="/static/banners/information/thencc.jpg"
       />
       <section>
         <PastSpeakersButton />
         <img
-          src="/static/speakers/2019/Jessica_Salisbury.png"
-          alt="Picture of Jessica Salisbury"
+          src="/static/speakers/2019/Ian_Johnson.png"
+          alt="Picture of Ian Johnson"
           className="speakerpic"
         />
         <div>
           <ul>
-            <b>Pronouns: </b> She/Her
+            <b>Pronouns: </b> He/Him
           </ul>
           <ul>
             <b>Talk Length: </b> 60 minute
           </ul>
         </div>
         <br />
-        <h2 tabIndex="0">The Power of the Feedback Loop</h2>
+        <h2 tabIndex="0">Reasonable Code</h2>
         <p tabIndex="0">
-          From how the body knows when to stop growing to saving lives
-          from speeding cars, the feedback loop is a mechanism that
-          secretly underlies every corner of our existence. The
-          concept spans schools of thought in psychology, economics,
-          biology and neuroscience and is now emerging in the
-          intersection of technology and behavioural science.
+          In a reasonable a system (i.e. a system that helps me to
+          understand it, to reason about it) I should be able to
+          understand how to make a change without holding the entire
+          system in my head. I should be able to reason where the
+          change needs to be made and reason about the impact it will
+          have.
         </p>
         <p tabIndex="0">
-          As technologists we are always looking to build products
-          that are engaging and effective, and as people we strive to
-          reach our full personal and working potential. Could the
-          feedback loop be the piece we&apos;re all missing?
-        </p>
-        <p tabIndex="0">
-          In this talk we&apos;ll dive into the structure of the
-          feedback loop, some impressive examples of how powerful they
-          can be and we&apos;ll explore how you can harness their
-          power in your work and personal life.
+          I want to explore what reasonable means to me, from the
+          processes of the team all the way down to an individual
+          block of code. Along the way, we will encounter existing
+          frameworks, tools, and patterns that our community has
+          developed over the years to help us to reason about our code
+          and processes; I feel that they have often been misused and
+          end up creating the opposite effect, adding unnecessary
+          complexity to how we work.
         </p>
       </section>
 
       <section>
-        <h2 tabIndex="0">About Jessica</h2>
+        <h2 tabIndex="0">About Ian</h2>
         <p tabIndex="0">
-          <b>
-            Behavioural Psychologist and Money Coaching Project
-            Manager at Tully.
-          </b>
+          <b>Software developer, sketch noter</b>
         </p>
         <p tabIndex="0">
-          Jess is a Behavioural Psychologist and the Project Manager
-          of Money Coaching at Tully, a business created to help
-          people who are worried about money. Jess leads on Tully’s
-          Money Coaching product which uses Open Banking data to
-          engage consumers in positive financial decision making to
-          improve their financial wellbeing. From helping those in
-          debt manage their money to providing early intervention for
-          those displaying detrimental patterns of behaviour, Jess is
-          using her academic background in psychology to create highly
-          personalised, data led Money Coaching that drives long term,
-          financial behavioural change for Tully customers.
+          Ian is a software developer working at Redgate, a company
+          that develops tools for developers and database
+          administrators.
         </p>
-      </section>
-
-      <section>
-        <h2 tabIndex="0">Video from 2019</h2>
-        <div className="video-frame">
-          <iframe
-            width="560"
-            height="315"
-            src="https://www.youtube.com/embed/bxz6LSOclXQ"
-            title="YouTube video player"
-            frameborder="0"
-            allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture"
-            allowfullscreen
-          ></iframe>
-        </div>
-        <div className="center-link">
-          <a
-            target="_blank"
-            rel="noreferrer"
-            aria-label="Jessica Salisbury: Youtube of Talk"
-            href="https://youtu.be/bxz6LSOclXQ"
-          >
-            Link to Youtube Recording
-          </a>
-        </div>
+        <p tabIndex="0">
+          Ian is passionate about writing maintainable code that
+          delivers on the needs of users. Though he considers himself
+          an introvert, Ian loves talking with other developers,
+          learning from their experiences and sharing his own.
+        </p>
+        <p tabIndex="0">
+          Outside of work, Ian is a passionate Star Wars fan and has
+          been known to make the occasional really bad pun, but all of
+          his code is "no-pun sourced" (sorry, couldn&apos;t resist).
+        </p>
       </section>
 
       <section>
@@ -117,20 +90,10 @@ export default function JessicaSalisbury() {
           <a
             target="_blank"
             rel="noreferrer"
-            aria-label="Jessica Salisbury: Twitter"
-            href="https://twitter.com/JessSalisburyy"
+            aria-label="Ian Johnson: Twitter"
+            href="https://twitter.com/ijohnson_tnf"
           >
             Twitter
-          </a>
-        </li>
-        <li>
-          <a
-            target="_blank"
-            rel="noreferrer"
-            aria-label="Jessica Salisbury: LinkedIn"
-            href="https://www.linkedin.com/in/jessica-salisbury-2a0920102/"
-          >
-            LinkedIn
           </a>
         </li>
       </section>
