+++ pages/pastconferences/pastspeakers/2019/joel-hammond-turner.js
@@ -14,75 +14,58 @@ import {
 
 library.add(faInstagram, faTwitter, faGithub, faLinkedin);
 
-export default function JessicaSalisbury() {
+export default function JoelHammondTurner() {
   return (
     <Layout>
       <Head>
-        <title>Jessica Salisbury</title>
+        <title>Joel Hammond-Turner</title>
       </Head>
       <Header
-        title={'Jessica Salisbury'}
+        title={'Joel Hammond-Turner'}
         banner="/static/banners/information/thencc.jpg"
       />
       <section>
         <PastSpeakersButton />
         <img
-          src="/static/speakers/2019/Jessica_Salisbury.png"
-          alt="Picture of Jessica Salisbury"
+          src="/static/speakers/2019/Joel_Hammond-Turner.png"
+          alt="Picture of Joel Hammond-Turner"
           className="speakerpic"
         />
         <div>
           <ul>
-            <b>Pronouns: </b> She/Her
+            <b>Pronouns: </b> He/Him
           </ul>
           <ul>
             <b>Talk Length: </b> 60 minute
           </ul>
         </div>
         <br />
-        <h2 tabIndex="0">The Power of the Feedback Loop</h2>
+        <h2 tabIndex="0">
+          You&apos;re the Tech Lead - *you* fix it!
+        </h2>
         <p tabIndex="0">
-          From how the body knows when to stop growing to saving lives
-          from speeding cars, the feedback loop is a mechanism that
-          secretly underlies every corner of our existence. The
-          concept spans schools of thought in psychology, economics,
-          biology and neuroscience and is now emerging in the
-          intersection of technology and behavioural science.
-        </p>
-        <p tabIndex="0">
-          As technologists we are always looking to build products
-          that are engaging and effective, and as people we strive to
-          reach our full personal and working potential. Could the
-          feedback loop be the piece we&apos;re all missing?
-        </p>
-        <p tabIndex="0">
-          In this talk we&apos;ll dive into the structure of the
-          feedback loop, some impressive examples of how powerful they
-          can be and we&apos;ll explore how you can harness their
-          power in your work and personal life.
+          Over the last couple of years, I&apos;ve presented 20 tricks
+          and tips that I&apos;ve found invaluable as a Tech Lead. But
+          in this session, I want to turn things around and look at
+          applying some of those to solving specific issues that are
+          common within many software development teams.
         </p>
       </section>
 
       <section>
-        <h2 tabIndex="0">About Jessica</h2>
+        <h2 tabIndex="0">About Joel</h2>
         <p tabIndex="0">
-          <b>
-            Behavioural Psychologist and Money Coaching Project
-            Manager at Tully.
-          </b>
+          <b>Tech Lead, Landmark Information Group</b>
         </p>
         <p tabIndex="0">
-          Jess is a Behavioural Psychologist and the Project Manager
-          of Money Coaching at Tully, a business created to help
-          people who are worried about money. Jess leads on Tully’s
-          Money Coaching product which uses Open Banking data to
-          engage consumers in positive financial decision making to
-          improve their financial wellbeing. From helping those in
-          debt manage their money to providing early intervention for
-          those displaying detrimental patterns of behaviour, Jess is
-          using her academic background in psychology to create highly
-          personalised, data led Money Coaching that drives long term,
-          financial behavioural change for Tully customers.
+          I&apos;ve a very broad experience of software development
+          over 15 years and a passion for both technology and elegance
+          in my solutions that make me an extremely capable software
+          architect.
+        </p>
+        <p tabIndex="0">
+          Personable and professional, I revel in complex challenges,
+          but always make time to coach and mentor team members.
         </p>
       </section>
 
@@ -92,7 +75,7 @@ export default function JessicaSalisbury() {
           <iframe
             width="560"
             height="315"
-            src="https://www.youtube.com/embed/bxz6LSOclXQ"
+            src="https://www.youtube.com/embed/bPHxRA_pE54"
             title="YouTube video player"
             frameborder="0"
             allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture"
@@ -103,8 +86,8 @@ export default function JessicaSalisbury() {
           <a
             target="_blank"
             rel="noreferrer"
-            aria-label="Jessica Salisbury: Youtube of Talk"
-            href="https://youtu.be/bxz6LSOclXQ"
+            aria-label="Joel Hammond-Turner: Youtube of Talk"
+            href="https://youtu.be/bPHxRA_pE54"
           >
             Link to Youtube Recording
           </a>
@@ -117,8 +100,8 @@ export default function JessicaSalisbury() {
           <a
             target="_blank"
             rel="noreferrer"
-            aria-label="Jessica Salisbury: Twitter"
-            href="https://twitter.com/JessSalisburyy"
+            aria-label="Joel Hammond-Turner: Twitter"
+            href="https://twitter.com/Rammesses"
           >
             Twitter
           </a>
@@ -127,8 +110,8 @@ export default function JessicaSalisbury() {
           <a
             target="_blank"
             rel="noreferrer"
-            aria-label="Jessica Salisbury: LinkedIn"
-            href="https://www.linkedin.com/in/jessica-salisbury-2a0920102/"
+            aria-label="Joel Hammond-Turner: LinkedIn"
+            href="https://www.linkedin.com/in/joelht/"
           >
             LinkedIn
           </a>
