+++ pages/pastconferences/pastspeakers/2019/neil-oconnor.js
@@ -14,75 +14,98 @@ import {
 
 library.add(faInstagram, faTwitter, faGithub, faLinkedin);
 
-export default function JessicaSalisbury() {
+export default function NeilOConnor() {
   return (
     <Layout>
       <Head>
-        <title>Jessica Salisbury</title>
+        <title>Neil O'Connor</title>
       </Head>
       <Header
-        title={'Jessica Salisbury'}
+        title={"Neil O'Connor"}
         banner="/static/banners/information/thencc.jpg"
       />
       <section>
         <PastSpeakersButton />
         <img
-          src="/static/speakers/2019/Jessica_Salisbury.png"
-          alt="Picture of Jessica Salisbury"
+          src="/static/speakers/2019/Neil_OConnor.png"
+          alt="Picture of Neil O'Connor"
           className="speakerpic"
         />
         <div>
           <ul>
-            <b>Pronouns: </b> She/Her
+            <b>Pronouns: </b> He/Him
           </ul>
           <ul>
             <b>Talk Length: </b> 60 minute
           </ul>
         </div>
         <br />
-        <h2 tabIndex="0">The Power of the Feedback Loop</h2>
+        <h2 tabIndex="0">
+          CTO secrets: How to get the best companies fighting to hire
+          you
+        </h2>
         <p tabIndex="0">
-          From how the body knows when to stop growing to saving lives
-          from speeding cars, the feedback loop is a mechanism that
-          secretly underlies every corner of our existence. The
-          concept spans schools of thought in psychology, economics,
-          biology and neuroscience and is now emerging in the
-          intersection of technology and behavioural science.
+          As a CTO, I spend a lot of my time hiring talent and
+          building high performance development teams. I have hired
+          hundreds of software professionals over the years, and I
+          still insist on interviewing every candidate personally. I
+          think I’ve got a pretty good hit rate: my teams have all had
+          great reputations in their local tech community, and I would
+          rehire - in a heartbeat - almost everyone I’ve hired in the
+          past.
         </p>
         <p tabIndex="0">
-          As technologists we are always looking to build products
-          that are engaging and effective, and as people we strive to
-          reach our full personal and working potential. Could the
-          feedback loop be the piece we&apos;re all missing?
+          It’s never been a better time to be working in the software
+          industry. Your skills are in demand, but that doesn’t mean
+          that everyone gets to work on the coolest tech in the most
+          forward thinking companies. How do you set yourself apart
+          from the crowd and get those companies falling over
+          themselves to hire you?
         </p>
         <p tabIndex="0">
-          In this talk we&apos;ll dive into the structure of the
-          feedback loop, some impressive examples of how powerful they
-          can be and we&apos;ll explore how you can harness their
-          power in your work and personal life.
+          In this session, I’ll share some of my secrets about what I
+          look for in candidates. What is it that makes a candidate
+          stand out? What do I look for in their CV? What is it about
+          how they approach the interview or the technical assessment
+          that impresses me? How important is formal education? What
+          do I look for in how the candidate spends their spare time?
+        </p>
+        <p tabIndex="0">
+          One myth to clear up right here: not everyone has to be a
+          “rockstar developer”. I look for a balanced mix of skills in
+          every candidate, and I’m often much more impressed by the
+          quiet one who has demonstrated great judgment, insight or a
+          capacity to learn. To continue the rockstar analogy, I’d
+          much rather hire the reliable drummer or the skilful backing
+          singer over the egotistical and unpredictable frontman!
+        </p>
+        <p tabIndex="0">
+          This session will offer clear, practical advice on how you
+          can maximise your personal value in the industry. Some of it
+          will be quick and easy to achieve, while some of it will
+          require you to make longer term changes to your mindset. But
+          whether you are an old hand eyeing your next career move, or
+          you are looking for your first job in the industry, there
+          will be something in this talk for you.
         </p>
       </section>
 
       <section>
-        <h2 tabIndex="0">About Jessica</h2>
+        <h2 tabIndex="0">About Neil</h2>
         <p tabIndex="0">
           <b>
-            Behavioural Psychologist and Money Coaching Project
-            Manager at Tully.
+            CTO at Koodoo, a Nottingham-based fintech in the Blenheim
+            Chalcot group
           </b>
         </p>
         <p tabIndex="0">
-          Jess is a Behavioural Psychologist and the Project Manager
-          of Money Coaching at Tully, a business created to help
-          people who are worried about money. Jess leads on Tully’s
-          Money Coaching product which uses Open Banking data to
-          engage consumers in positive financial decision making to
-          improve their financial wellbeing. From helping those in
-          debt manage their money to providing early intervention for
-          those displaying detrimental patterns of behaviour, Jess is
-          using her academic background in psychology to create highly
-          personalised, data led Money Coaching that drives long term,
-          financial behavioural change for Tully customers.
+          Neil has been building forward-thinking development teams
+          since 1862. He was founding CTO at local companies Koodoo
+          and Oakbrook Finance, and has worked in a wide array of
+          techie roles in numerous industries. He believes that if
+          something is worth doing, it&apos;s worth doing right, and
+          aims to create environments where development teams can
+          excel at what they do best.
         </p>
       </section>
 
@@ -92,7 +115,7 @@ export default function JessicaSalisbury() {
           <iframe
             width="560"
             height="315"
-            src="https://www.youtube.com/embed/bxz6LSOclXQ"
+            src="https://www.youtube.com/embed/FHhgTXiv-Rw"
             title="YouTube video player"
             frameborder="0"
             allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture"
@@ -103,8 +126,8 @@ export default function JessicaSalisbury() {
           <a
             target="_blank"
             rel="noreferrer"
-            aria-label="Jessica Salisbury: Youtube of Talk"
-            href="https://youtu.be/bxz6LSOclXQ"
+            aria-label="Neil: Youtube of Talk"
+            href="https://youtu.be/FHhgTXiv-Rw"
           >
             Link to Youtube Recording
           </a>
@@ -117,20 +140,10 @@ export default function JessicaSalisbury() {
           <a
             target="_blank"
             rel="noreferrer"
-            aria-label="Jessica Salisbury: Twitter"
-            href="https://twitter.com/JessSalisburyy"
+            aria-label="Neil: Twitter"
+            href="https://twitter.com/1stmanonthesun"
           >
             Twitter
-          </a>
-        </li>
-        <li>
-          <a
-            target="_blank"
-            rel="noreferrer"
-            aria-label="Jessica Salisbury: LinkedIn"
-            href="https://www.linkedin.com/in/jessica-salisbury-2a0920102/"
-          >
-            LinkedIn
           </a>
         </li>
       </section>
