+++ pages/pastconferences/pastspeakers/2019/zac-braddy.js
@@ -14,75 +14,97 @@ import {
 
 library.add(faInstagram, faTwitter, faGithub, faLinkedin);
 
-export default function JessicaSalisbury() {
+export default function ZacBraddy() {
   return (
     <Layout>
       <Head>
-        <title>Jessica Salisbury</title>
+        <title>Zac Braddy</title>
       </Head>
       <Header
-        title={'Jessica Salisbury'}
+        title={'Zac Braddy'}
         banner="/static/banners/information/thencc.jpg"
       />
       <section>
         <PastSpeakersButton />
         <img
-          src="/static/speakers/2019/Jessica_Salisbury.png"
-          alt="Picture of Jessica Salisbury"
+          src="/static/speakers/2019/Zac_Braddy.png"
+          alt="Picture of Zac Braddy"
           className="speakerpic"
         />
         <div>
           <ul>
-            <b>Pronouns: </b> She/Her
+            <b>Pronouns: </b> He/Him
           </ul>
           <ul>
             <b>Talk Length: </b> 60 minute
           </ul>
         </div>
         <br />
-        <h2 tabIndex="0">The Power of the Feedback Loop</h2>
+        <h2 tabIndex="0">
+          All the mistakes I&apos;ve made trying to implement
+          Microservices
+        </h2>
         <p tabIndex="0">
-          From how the body knows when to stop growing to saving lives
-          from speeding cars, the feedback loop is a mechanism that
-          secretly underlies every corner of our existence. The
-          concept spans schools of thought in psychology, economics,
-          biology and neuroscience and is now emerging in the
-          intersection of technology and behavioural science.
+          We&apos;re now entering the next age of Microservices. For
+          about a decade now we&apos;ve been doing battle with the
+          pointy haired bosses of the world to convince them of what
+          seems so painfully obvious to us; that there are many varied
+          virtues of microservices architecture.
         </p>
         <p tabIndex="0">
-          As technologists we are always looking to build products
-          that are engaging and effective, and as people we strive to
-          reach our full personal and working potential. Could the
-          feedback loop be the piece we&apos;re all missing?
+          Now that they all have Netflix accounts and they&apos;ve
+          seen the benefits first hand we now seem to have those same
+          bosses bursting into the office waving their trade magazine
+          and asking us if we&apos;ve heard about this "micro-system"
+          thing. Flustered, we&apos;re having coming to terms now with
+          how we actually deliver on our lofty promises without
+          Amazon&apos;s finances and resources.
         </p>
         <p tabIndex="0">
-          In this talk we&apos;ll dive into the structure of the
-          feedback loop, some impressive examples of how powerful they
-          can be and we&apos;ll explore how you can harness their
-          power in your work and personal life.
+          Some people have been getting it right the first time
+          they&apos;ve tried which is great! Other people, like me,
+          have got it right as well.....but....not before getting it
+          wrong repeatedly for oh so many different reasons! So, I
+          guess that makes me an expert, right?!
+        </p>
+        <p tabIndex="0">
+          Come listen as I shamefully regale you with all ways in
+          which I&apos;ve ham-fisted the most elegant architecture
+          currently known to humanity in hopes that I might save you
+          from the same sleepless nights.
         </p>
       </section>
 
       <section>
-        <h2 tabIndex="0">About Jessica</h2>
+        <h2 tabIndex="0">About Zac</h2>
         <p tabIndex="0">
-          <b>
-            Behavioural Psychologist and Money Coaching Project
-            Manager at Tully.
-          </b>
+          <b>Lead Developer</b>
         </p>
         <p tabIndex="0">
-          Jess is a Behavioural Psychologist and the Project Manager
-          of Money Coaching at Tully, a business created to help
-          people who are worried about money. Jess leads on Tully’s
-          Money Coaching product which uses Open Banking data to
-          engage consumers in positive financial decision making to
-          improve their financial wellbeing. From helping those in
-          debt manage their money to providing early intervention for
-          those displaying detrimental patterns of behaviour, Jess is
-          using her academic background in psychology to create highly
-          personalised, data led Money Coaching that drives long term,
-          financial behavioural change for Tully customers.
+          Zac is a Lead Fullstack Javascript developer at Koodoo.io
+          where he helps to try and save people money on their
+          mortgages. This makes him feel super good about how he
+          spends his days.
+        </p>
+        <p tabIndex="0">
+          Zac has been developing for a number of years, 5 or 6 I
+          think, I don&apos;t know who is counting? In that time has
+          seen a very diverse set of tech and industries. Up until his
+          current role he was working with the .NET stack so even
+          languages are sacred to him. For Zac it&apos;s all about the
+          next challenge and in the past that thirst for knowledge has
+          seen him working with everything from hulking 20 year old
+          Classic ASP monoliths to greenfield node microservices on
+          bleed edge architecture and good amount of stuff in between.
+        </p>
+        <p tabIndex="0">
+          On the side Zac is active in the developer community writing
+          blog posts, doing talks and mentoring other developers
+          through their journeys in meet ups and online. When he
+          isn&apos;t in front of his computer at work he is out
+          geocaching with his sons and wife, playing D and D with
+          friends or......let&apos;s face it probably on the computer
+          still coding or gaming.
         </p>
       </section>
 
@@ -92,7 +114,7 @@ export default function JessicaSalisbury() {
           <iframe
             width="560"
             height="315"
-            src="https://www.youtube.com/embed/bxz6LSOclXQ"
+            src="https://www.youtube.com/embed/43qsKWUNUpc"
             title="YouTube video player"
             frameborder="0"
             allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture"
@@ -103,8 +125,8 @@ export default function JessicaSalisbury() {
           <a
             target="_blank"
             rel="noreferrer"
-            aria-label="Jessica Salisbury: Youtube of Talk"
-            href="https://youtu.be/bxz6LSOclXQ"
+            aria-label="Zac: Youtube of Talk"
+            href="https://youtu.be/43qsKWUNUpc"
           >
             Link to Youtube Recording
           </a>
@@ -117,8 +139,8 @@ export default function JessicaSalisbury() {
           <a
             target="_blank"
             rel="noreferrer"
-            aria-label="Jessica Salisbury: Twitter"
-            href="https://twitter.com/JessSalisburyy"
+            aria-label="Zac Braddy: Twitter"
+            href="https://twitter.com/ZackerTheHacker"
           >
             Twitter
           </a>
@@ -127,8 +149,8 @@ export default function JessicaSalisbury() {
           <a
             target="_blank"
             rel="noreferrer"
-            aria-label="Jessica Salisbury: LinkedIn"
-            href="https://www.linkedin.com/in/jessica-salisbury-2a0920102/"
+            aria-label="Zac Braddy: LinkedIn"
+            href="https://www.linkedin.com/in/zac-braddy-17a81b22/"
           >
             LinkedIn
           </a>
